backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
beautifulsoup4==4.12.3
duckduckgo-search==6.3.7
groq==0.12.0
python-dotenv==1.0.1
pydantic==2.9.2
